requirements.txt
pytest==8.2.0
mimesis==16.0.0
requests==2.31.0
allure-pytest==2.13.5